Move addSplitter to network_tools
* Move isRectangular, Wiring, Config, Configs  to network_tools
* Move assert, algorithm includes to Network tools
* Add network_tool header defs for Wiring, Config, Configs, addSplitter, is_rectangular
* Fix Us Wiring type addSplitter header declaration
* Improve network tools header documentation
* Drop transpose from existsBalancer header
