+++ lib/network_tools.hpp
@@ -4,6 +4,10 @@
 
 #include "types.hpp"
 
+//////////////////////////////
+// Matrix operations
+//////////////////////////////
+
 // Generate a row of zeros
 Row zeroRow(int size);
 
@@ -19,11 +23,26 @@ Row rowMultiply(Row row, double multiplier);
 // Generate an identity matrix matrix
 Matrix identityMatrix(int size);
 
+bool isRectangular(Matrix network);
+
 // Extract a column from a matrix
 Row getColumn(Matrix matrix, int column_position);
 
-// Transpos a matrix
+// Transpose a matrix
 Matrix transpose(Matrix matrix);
+
+//////////////////////////////
+// Network operations
+//////////////////////////////
+
+// Represents a splitter's connection to an existing network
+using Wiring = vector<int>;
+
+// A Pair of wirings for a splitter's input and output
+using Config = vector<Wiring>;
+
+// A collection of wirings buildin a network
+using Configs = vector<Config>;
 
 // Generate an empty network
 Network emptyNetwork(int size);
@@ -33,3 +52,10 @@ void link(Network& nodes, int source, int target);
 
 // Find a node's index in a network
 int nodeNum(Network nodes, Node* node);
+
+// network[i] is the flow from inputs to i-th output.
+// splitter_inputs has an entry of -1 for a new input that's not already part of
+// the network. splitter_outputs has an entry of -1 for a new output There must
+// be at least one output Network must have at least one input
+Matrix addSplitter(Matrix network, Wiring splitter_inputs,
+                   Wiring splitter_outputs);
